+++ Sources/Plasma/PubUtilLib/plNetClient/plNetClientMgrSend.cpp
@@ -98,14 +98,14 @@ int plNetClientMgr::ISendDirtyState(double secs)
 {
     std::vector<plSynchedObject::StateDefn> carryOvers;
 
-    int32_t num=plSynchedObject::GetNumDirtyStates();
+    size_t num=plSynchedObject::GetNumDirtyStates();
 #if 0
     if (num)
     {
         DebugMsg("{} dirty sdl state msgs queued, t={f}", num, secs);
     }
 #endif
-    int32_t i;
+    size_t i;
     for(i=0;i<num;i++)
     {
         plSynchedObject::StateDefn* state=plSynchedObject::GetDirtyState(i);
